create a method for full section generation
